backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pdfplumber==0.11.4
PyMuPDF
anthropic==0.43.0
python-dotenv==1.0.1
edge-tts
httpx
firebase-admin
